refactor: update package structure from com.tak.chatws to com.tak.chat_ws
