+++ chat-ws/src/main/java/com/tak/chat_ws/chat/ChatController.java
@@ -1,6 +1,6 @@
-package com.tak.chatws.chat;
+package com.tak.chat_ws.chat;
 
-import com.tak.chatws.redis.RedisPublisher;
+import com.tak.chat_ws.redis.RedisPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
@@ -10,13 +10,9 @@ import org.springframework.stereotype.Controller;
 @Controller
 @RequiredArgsConstructor
 public class ChatController {
-    private final SimpMessagingTemplate template;
     private final RedisPublisher redisPublisher;
     @MessageMapping("/chat/{roomId}")
     public void sendMessage(@DestinationVariable String roomId, ChatMessageDto chatMessageDto) {
-        // 메시지 처리 로직
-//        System.out.println("ChatMessageDto = " + chatMessageDto);
         redisPublisher.publish(chatMessageDto);
-//        template.convertAndSend("/topic/" + roomId, chatMessageDto);
     }
 }
