+++ chat-ws/src/main/java/com/tak/chat_ws/chat/ChatMessageDto.java
@@ -1,4 +1,4 @@
-package com.tak.chatws.chat;
+package com.tak.chat_ws.chat;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
