+++ chat-ws/src/main/java/com/tak/chat_ws/redis/PubSubTestController.java
@@ -1,6 +1,6 @@
-package com.tak.chatws.redis;
+package com.tak.chat_ws.redis;
 
-import com.tak.chatws.chat.ChatMessageDto;
+import com.tak.chat_ws.chat.ChatMessageDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
