+++ chat-ws/src/test/java/com/tak/chat_ws/ChatWsApplicationTests.java
@@ -1,4 +1,4 @@
-package com.tak.chatws;
+package com.tak.chat_ws;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
